Grant Crystalis when appropriate
Since Crystalis is handled the same as any other item, it turns out my ASM changes to the ROM turned Mesia at the top of tower into an APItem location. So, this change fixes that and grants Crystalis.  This could theoretically be an item the multiworld is actually aware of, but it breaks the game if granted early, so I'm not letting the server know about it.