Add Mega Man 2 color support
Each item gets a color pair depending on the palette it uses in the inventory screen, labeled by element and whether the palette is used on the equipment screen or the consumable screen. Yes, this does result in WaterEquip and WaterConsumable being identical (and actually the same blues used as Mega Man's default color), but I prefer to keep them separated like this in case I change my mind on one of those two categories later.