Replace Accessories when Vb is on
When Vb (Don't buff bonus items) is on, replace Hazmat Suit with Gas Mask and Speed Boots with Leather Shoes.  The logic already accounts for the different items correctly, but the items were never created prior to this change. This could have resulted in a seed that expects you to use Leather Boots as though they were Speed Boots (e.g. climbing the slope to the Tornado Bracelet spot) or Gas Mask as though it were Hazmat Suit (e.g. for crossing spike pits).