Add missing rule to the fisherman when vanilla dolphin is on
Fisherman doesn't spawn with vanilla dolphin unless you have both his trade-in and the shell flute and you've healed the dolphin.  This didn't matter before ER was implemented, because the only thing Boat Access was useful for was activating the shell flute (which also had a shell flute requirement), but that might not be the case once the Beach House doesn't always contain the Beach House.